requirements.txt
langchain
langgraph
langchain-google-genai
langchain-community
langchain-tavily
python-dotenv
requests
ipython